renaming to routeHandler
