+++ package.json
@@ -37,13 +37,13 @@
     "@jns42/jns42-schema-oas-v3-0": "^0.1.4",
     "@jns42/jns42-schema-oas-v3-1": "^0.1.4",
     "@jns42/jns42-schema-swagger-v2": "^0.1.6",
-    "@oa42/oa42-lib": "^0.1.1",
-    "@types/node": "^20.8.8",
+    "@oa42/oa42-lib": "^0.2.0",
+    "@types/node": "^20.8.9",
     "@types/yargs": "^17.0.29",
     "camelcase": "^8.0.0",
     "goodrouter": "^2.0.7",
     "tslib": "^2.6.2",
-    "type-fest": "^4.5.0",
+    "type-fest": "^4.6.0",
     "typescript": "^5.2.2",
     "yaml": "^2.3.3",
     "yargs": "^17.7.2"
