+++ src/generators/functions/transform-incoming-request.ts
@@ -78,6 +78,297 @@ export class TransformIncomingRequestCodeGenerator extends CodeGeneratorBase {
   ) {
     const { factory: f } = this;
 
-    yield f.createThrowStatement(f.createStringLiteral("TODO"));
+    const operationIncomingParametersName = toPascal(
+      operationModel.name,
+      "request",
+      "parameters",
+    );
+
+    /**
+     * read some headers
+     */
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestCookieHeader"),
+            undefined,
+            undefined,
+            f.createCallExpression(
+              f.createPropertyAccessExpression(
+                f.createIdentifier("lib"),
+                f.createIdentifier("getParameterValue"),
+              ),
+              undefined,
+              [
+                f.createPropertyAccessExpression(
+                  f.createIdentifier("incomingRequest"),
+                  f.createIdentifier("headers"),
+                ),
+                f.createStringLiteral("cookie"),
+              ],
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestAcceptHeader"),
+            undefined,
+            undefined,
+            f.createCallExpression(
+              f.createPropertyAccessExpression(
+                f.createIdentifier("lib"),
+                f.createIdentifier("getParameterValue"),
+              ),
+              undefined,
+              [
+                f.createPropertyAccessExpression(
+                  f.createIdentifier("incomingRequest"),
+                  f.createIdentifier("headers"),
+                ),
+                f.createStringLiteral("accept"),
+              ],
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestContentTypeHeader"),
+            undefined,
+            undefined,
+            f.createCallExpression(
+              f.createPropertyAccessExpression(
+                f.createIdentifier("lib"),
+                f.createIdentifier("getParameterValue"),
+              ),
+              undefined,
+              [
+                f.createPropertyAccessExpression(
+                  f.createIdentifier("incomingRequest"),
+                  f.createIdentifier("headers"),
+                ),
+                f.createStringLiteral("content-type"),
+              ],
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    /**
+     * now we put the raw parameters in variables, path parameters are already
+     * present, they are in the methods arguments
+     */
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestHeaders"),
+            undefined,
+            undefined,
+            f.createPropertyAccessExpression(
+              f.createIdentifier("incomingRequest"),
+              f.createIdentifier("headers"),
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestQuery"),
+            undefined,
+            undefined,
+            f.createCallExpression(
+              f.createPropertyAccessExpression(
+                f.createIdentifier("lib"),
+                f.createIdentifier("parseParameters"),
+              ),
+              undefined,
+              [
+                f.createPropertyAccessExpression(
+                  f.createIdentifier("incomingRequest"),
+                  f.createIdentifier("query"),
+                ),
+                f.createStringLiteral("&"),
+                f.createStringLiteral("="),
+              ],
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestCookie"),
+            undefined,
+            undefined,
+            f.createCallExpression(
+              f.createPropertyAccessExpression(
+                f.createIdentifier("lib"),
+                f.createIdentifier("parseParameters"),
+              ),
+              undefined,
+              [
+                f.createBinaryExpression(
+                  f.createIdentifier("requestCookieHeader"),
+                  f.createToken(ts.SyntaxKind.QuestionQuestionToken),
+                  f.createStringLiteral(""),
+                ),
+                f.createStringLiteral("; "),
+                f.createStringLiteral("="),
+              ],
+            ),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    /**
+     * create the request parameters object
+     */
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("requestParameters"),
+            undefined,
+            f.createTypeReferenceNode(
+              f.createQualifiedName(
+                f.createIdentifier("shared"),
+                f.createIdentifier(operationIncomingParametersName),
+              ),
+            ),
+            undefined,
+          ),
+        ],
+        ts.NodeFlags.Let,
+      ),
+    );
+
+    yield f.createTryStatement(
+      f.createBlock(
+        [
+          f.createExpressionStatement(
+            f.createBinaryExpression(
+              f.createIdentifier("requestParameters"),
+              f.createToken(ts.SyntaxKind.EqualsToken),
+              f.createObjectLiteralExpression(
+                [
+                  ...operationModel.pathParameters.map((parameterModel) =>
+                    f.createPropertyAssignment(
+                      toCamel(parameterModel.name),
+                      f.createElementAccessExpression(
+                        f.createIdentifier("routeParameters"),
+                        f.createStringLiteral(parameterModel.name),
+                      ),
+                    ),
+                  ),
+                  ...operationModel.headerParameters.map((parameterModel) =>
+                    f.createPropertyAssignment(
+                      toCamel(parameterModel.name),
+                      f.createElementAccessExpression(
+                        f.createIdentifier("requestHeaders"),
+                        f.createStringLiteral(parameterModel.name),
+                      ),
+                    ),
+                  ),
+                  ...operationModel.queryParameters.map((parameterModel) =>
+                    f.createPropertyAssignment(
+                      toCamel(parameterModel.name),
+                      f.createElementAccessExpression(
+                        f.createIdentifier("requestQuery"),
+                        f.createStringLiteral(parameterModel.name),
+                      ),
+                    ),
+                  ),
+                  ...operationModel.cookieParameters.map((parameterModel) =>
+                    f.createPropertyAssignment(
+                      toCamel(parameterModel.name),
+                      f.createElementAccessExpression(
+                        f.createIdentifier("requestCookie"),
+                        f.createStringLiteral(parameterModel.name),
+                      ),
+                    ),
+                  ),
+                ],
+                true,
+              ),
+            ),
+          ),
+        ],
+        true,
+      ),
+      f.createCatchClause(
+        undefined,
+        f.createBlock(
+          [
+            f.createThrowStatement(
+              f.createNewExpression(f.createIdentifier("Error"), undefined, [
+                f.createStringLiteral("parameter error"),
+              ]),
+            ),
+          ],
+          true,
+        ),
+      ),
+      undefined,
+    );
+
+    /**
+     * now lets construct the incoming request object, this object will be
+     * passed to the operation handler later
+     */
+
+    yield f.createVariableStatement(
+      undefined,
+      f.createVariableDeclarationList(
+        [
+          f.createVariableDeclaration(
+            f.createIdentifier("incomingOperationRequest"),
+            undefined,
+            undefined,
+            f.createNull(),
+          ),
+        ],
+        ts.NodeFlags.Const,
+      ),
+    );
+
+    yield f.createReturnStatement(
+      f.createIdentifier("incomingOperationRequest"),
+    );
   }
 }
